requirements.txt
Pillow==7.2.0
pandas==1.1.0
pyinstaller